do not use sessions when running pipelines

+++ ContainerHandling/Remove-NavContainerSession.ps1
@@ -19,18 +19,17 @@ function Remove-BcContainerSession {
     Process {
         if ($sessions.ContainsKey($containerName)) {
             $session = $sessions[$containerName]
-            try {
+            if ($bcContainerHelperConfig.KillPsSessionProcess) {
                 $inspect = docker inspect $containerName | ConvertFrom-Json
                 if ($inspect.HostConfig.Isolation -eq "process") {
                     $processID = Invoke-Command -Session $session -ScriptBlock { $PID }
                     Stop-Process -Id $processID -Force
                 }
                 else {
-
                     Remove-PSSession -Session $session
                 }
             }
-            catch {
+            else {
                 Remove-PSSession -Session $session
             }
